# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flipkart Clone API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  imageUrl?: string | null;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  price: number;
  originalPrice?: number | null;
  discountPercent?: number | null;
  rating?: number | null;
  reviewCount?: number | null;
  stock: number;
  categoryId: number;
  categoryName?: string | null;
  imageUrl?: string | null;
  brand?: string | null;
  isFeatured: boolean;
}

export type ProductDetailSpecifications = { [key: string]: string } | null;

export type ProductDetail = Product & {
  images?: string[];
  specifications?: ProductDetailSpecifications;
  highlights?: string[] | null;
};

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  product: Product;
}

export interface CartResponse {
  items: CartItem[];
  subtotal: number;
  deliveryCharge: number;
  total: number;
  itemCount: number;
}

export interface AddToCartRequest {
  productId: number;
  quantity?: number;
}

export interface UpdateCartItemRequest {
  quantity: number;
}

export interface Address {
  fullName: string;
  phone: string;
  addressLine1: string;
  addressLine2?: string | null;
  city: string;
  state: string;
  pincode: string;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImage?: string | null;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  status: string;
  subtotal: number;
  deliveryCharge: number;
  total: number;
  shippingAddress: Address;
  items: OrderItem[];
  createdAt: string;
  estimatedDelivery?: string | null;
}

export interface PlaceOrderRequest {
  shippingAddress: Address;
  paymentMethod: string;
}

export type ListProductsParams = {
  /**
   * Search by product name
   */
  search?: string;
  /**
   * Filter by category ID
   */
  categoryId?: number;
  /**
   * Minimum price filter
   */
  minPrice?: number;
  /**
   * Maximum price filter
   */
  maxPrice?: number;
  /**
   * Page number (default 1)
   */
  page?: number;
  /**
   * Items per page (default 20)
   */
  limit?: number;
};