# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Flipkart Clone API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      description: Returns a list of products with optional filtering
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by product name
        - name: categoryId
          in: query
          schema:
            type: number
          description: Filter by category ID
        - name: minPrice
          in: query
          schema:
            type: number
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: number
          description: Maximum price filter
        - name: page
          in: query
          schema:
            type: number
          description: Page number (default 1)
        - name: limit
          in: query
          schema:
            type: number
          description: Items per page (default 20)
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"

  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "200":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /cart/items/{id}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
        "404":
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "201":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request (empty cart, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    Category:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          nullable: true
      required:
        - id
        - name
        - slug

    Product:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        discountPercent:
          type: number
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: number
          nullable: true
        stock:
          type: number
        categoryId:
          type: number
        categoryName:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        isFeatured:
          type: boolean
      required:
        - id
        - name
        - slug
        - price
        - stock
        - categoryId
        - isFeatured

    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            images:
              type: array
              items:
                type: string
            specifications:
              type: object
              additionalProperties:
                type: string
              nullable: true
            highlights:
              type: array
              items:
                type: string
              nullable: true

    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
      required:
        - products
        - total
        - page
        - limit
        - totalPages

    CartItem:
      type: object
      properties:
        id:
          type: number
        productId:
          type: number
        quantity:
          type: number
        product:
          $ref: "#/components/schemas/Product"
      required:
        - id
        - productId
        - quantity
        - product

    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        deliveryCharge:
          type: number
        total:
          type: number
        itemCount:
          type: number
      required:
        - items
        - subtotal
        - deliveryCharge
        - total
        - itemCount

    AddToCartRequest:
      type: object
      properties:
        productId:
          type: number
        quantity:
          type: number
      required:
        - productId

    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: number
      required:
        - quantity

    Address:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
      required:
        - fullName
        - phone
        - addressLine1
        - city
        - state
        - pincode

    OrderItem:
      type: object
      properties:
        id:
          type: number
        productId:
          type: number
        productName:
          type: string
        productImage:
          type: string
          nullable: true
        quantity:
          type: number
        price:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - price

    Order:
      type: object
      properties:
        id:
          type: number
        status:
          type: string
        subtotal:
          type: number
        deliveryCharge:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        estimatedDelivery:
          type: string
          nullable: true
      required:
        - id
        - status
        - subtotal
        - deliveryCharge
        - total
        - shippingAddress
        - items
        - createdAt

    PlaceOrderRequest:
      type: object
      properties:
        shippingAddress:
          $ref: "#/components/schemas/Address"
        paymentMethod:
          type: string
      required:
        - shippingAddress
        - paymentMethod