# lib/api-zod/src/generated/types/addToCartRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flipkart Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface AddToCartRequest {
  productId: number;
  quantity?: number;
}