# lib/api-zod/src/generated/types/cartResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flipkart Clone API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";

export interface CartResponse {
  items: CartItem[];
  subtotal: number;
  deliveryCharge: number;
  total: number;
  itemCount: number;
}