# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flipkart Clone API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from "./address";
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  status: string;
  subtotal: number;
  deliveryCharge: number;
  total: number;
  shippingAddress: Address;
  items: OrderItem[];
  createdAt: string;
  estimatedDelivery?: string | null;
}