# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flipkart Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImage?: string | null;
  quantity: number;
  price: number;
}