# lib/api-zod/src/generated/types/placeOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flipkart Clone API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from "./address";

export interface PlaceOrderRequest {
  shippingAddress: Address;
  paymentMethod: string;
}